refactor: ui folder
-Reorganize ui folder for improved readability/organization.
